refactor(core): point external call sites at web_chat (#5002)
Harness-family consumers (core::{socketio,jsonrpc,all,observability,
event_bus}, agent::task_dispatcher, threads::ops, agentbox::invoker,
cron::scheduler, flows::ops, test_support::introspect) now import the
conduit from openhuman::web_chat.

src/core/all.rs keeps its DomainGroup::Channels tag: the web RPC surface
must stay hidden by DomainSet::harness() at runtime exactly as today.
Re-tagging it would be an unreviewed behaviour change; #4801 will cfg
this registration site instead.

rustfmt reflows 4 call sites purely because the shorter path now fits.
