+++ src/core/jsonrpc.rs
@@ -1479,7 +1479,7 @@ async fn events_handler(
     }
 
     let client_id = query.client_id;
-    let rx = crate::openhuman::channels::providers::web::subscribe_web_channel_events();
+    let rx = crate::openhuman::web_chat::subscribe_web_channel_events();
     let stream = tokio_stream::wrappers::BroadcastStream::new(rx).filter_map(
         move |item| -> Option<Result<Event, std::convert::Infallible>> {
             let event = match item {
@@ -2350,18 +2350,18 @@ pub async fn bootstrap_core_runtime(
     // `start_channels` is skipped for web-chat-only cores. Without this an
     // unguarded standalone/CLI/Docker core would park a plan review that never
     // reaches the UI and dies at the gate TTL. Idempotent (Once-guarded).
-    crate::openhuman::channels::providers::web::register_approval_surface_subscriber();
+    crate::openhuman::web_chat::register_approval_surface_subscriber();
     // Bridge emergency-stop halt/resume → the `automation_halt` broadcast on the
     // same always-run boot path. `start_channels` (which also registers this)
     // is skipped on a web-chat-only desktop with no listening integrations, so
     // without this a halt/resume initiated from the CLI or another client would
     // never reach the UI. Idempotent (Once-guarded). (#4255)
-    crate::openhuman::channels::providers::web::register_automation_halt_subscriber();
+    crate::openhuman::web_chat::register_automation_halt_subscriber();
     // Egress-surface bridge (privacy epic S2, #4436) — registered
     // unconditionally alongside the approval surface so external-transfer
     // disclosures reach the UI even on cores that skip `start_channels` or run
     // with the approval gate disabled. Idempotent (OnceLock-guarded).
-    crate::openhuman::channels::providers::web::register_egress_surface_subscriber();
+    crate::openhuman::web_chat::register_egress_surface_subscriber();
 
     if decision.install_gate {
         // Per-launch correlation token for the approval gate. This is
@@ -2382,7 +2382,7 @@ pub async fn bootstrap_core_runtime(
         );
         // (The approval/plan-review surface bridge is registered unconditionally
         // above — it must run even when this gate-install branch is skipped.)
-        crate::openhuman::channels::providers::web::register_artifact_surface_subscriber();
+        crate::openhuman::web_chat::register_artifact_surface_subscriber();
     } else {
         log::error!(
             "[runtime] approval gate DISABLED (OPENHUMAN_APPROVAL_GATE=0 honored on host={}) — \
@@ -2402,7 +2402,7 @@ pub async fn bootstrap_core_runtime(
     // `if approval_gate` block so artifact events still publish when the user
     // sets OPENHUMAN_APPROVAL_GATE=0 (CR #3328947323 on PR #3026). Idempotent
     // (OnceLock-guarded inside register_artifact_surface_subscriber).
-    crate::openhuman::channels::providers::web::register_artifact_surface_subscriber();
+    crate::openhuman::web_chat::register_artifact_surface_subscriber();
 
     // --- Workspace migrations --------------------------------------------
     crate::openhuman::startup::run_workspace_migrations(&workspace_dir);
