+++ src/openhuman/agent/turn_origin.rs
@@ -4,7 +4,7 @@
 //! consistent decisions across web, channel, subconscious, and cron entry
 //! points without relying on the *absence* of other task-locals as a signal.
 //!
-//! Every entry point that drives the agent loop ([`crate::openhuman::channels::providers::web`],
+//! Every entry point that drives the agent loop ([`crate::openhuman::web_chat`],
 //! [`crate::openhuman::channels::runtime::dispatch`], [`crate::openhuman::subconscious`],
 //! [`crate::openhuman::cron`], CLI) MUST scope a real [`AgentTurnOrigin`]
 //! around its `run_turn` invocation. Any path that fails to do so is treated
