+++ src/openhuman/test_support/README.md
@@ -56,7 +56,7 @@ This module owns no state of its own — it mutates/reads state owned by other d
 - `crate::openhuman::config` — `Config::load_or_init`/`save`, `workspace_dir`, `clear_active_user`, `default_root_openhuman_dir`; config is the source of truth for the workspace root and the fields reset wipes.
 - `crate::openhuman::cron` — `clear_all_jobs` to wipe scheduled jobs during reset.
 - `crate::openhuman::memory::read_rpc` — `wipe_all_rpc` to clear memory-tree rows, content dirs, and sync state.
-- `crate::openhuman::channels::providers::web` — `in_flight_entries_for_test` to snapshot the live `IN_FLIGHT` chat map.
+- `crate::openhuman::web_chat` — `in_flight_entries_for_test` to snapshot the live `IN_FLIGHT` chat map.
 - `crate::openhuman::wallet` — `prepared_quotes_for_test` and `PreparedTransaction` to snapshot prepared quotes.
 - `crate::core::all` — `ControllerFuture`, `RegisteredController` for handler wiring.
 - `crate::core::{ControllerSchema, FieldSchema, TypeSchema}` — controller schema types.
