+++ src/openhuman/web_chat/ops.rs
@@ -728,7 +728,7 @@ pub async fn start_chat(
 
         match result {
             Ok(chat_result) => {
-                crate::openhuman::channels::providers::web::presentation::deliver_response(
+                crate::openhuman::web_chat::presentation::deliver_response(
                     &client_id_task,
                     &thread_id_task,
                     &request_id_task,
@@ -931,7 +931,7 @@ async fn spawn_parallel_turn(
 
         match result {
             Some(Ok(chat_result)) => {
-                crate::openhuman::channels::providers::web::presentation::deliver_response(
+                crate::openhuman::web_chat::presentation::deliver_response(
                     &client_id_task,
                     &thread_id_task,
                     &request_id_task,
